AD-2: implement RTCPeerConnection feature
Added a `peerConnection` library to facilitate `RTCPeerConnection`s.
Only data channels are created. Peer connections are managed by the
aptly name `peerManager` class.

The `wsTransport` class is a wrapper for communications with the
AWS API Gateway + Lambda services.

A couple of SVGs are thrown in to display connecting and connected
status between peers.